cat-vs-dog-webapp/requirements.txt
Flask==1.1.2
numpy==1.19.1
Pillow==7.2.0
torch==1.1.0
torchvision==0.3.0

